requirements.txt
matplotlib==3.7.2
numpy==1.24.3
pandas==1.5.3
scikit_learn==1.2.2
tensorflow==2.13.0
yfinance==0.2.28
polygon-api-client==1.12.3
python-dotenv==1.0.0
sqlite3
bs4

